requirements.txt
streamlit>=1.28.0
qdrant-client>=1.6.0
google-generativeai>=0.3.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
docling>=1.0.0
rank-bm25>=0.2.2
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
rouge-score>=0.1.2
sacrebleu>=2.3.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
plotly>=5.15.0
